Add Privacy and Security Section (#168)
Merging this editorial change to allow the horizontal reviews to be raised. If there are further issues, please raise them separately.

Closes #166.

* Add Privacy and Security Section
* Add fingerprinting images

Address review feedback. The images themselves are slightly cryptic and don't have any link back to something that explains why they're there, so I'm not sure about the overall benefit.

* Split Privacy and Security Considerations into two

to satisfy the horizontal review requirements of the groups that review for privacy and security.

* Add note re avoiding exposing audio resources by embedding them