Split Script Event Description out as a separate object
Closes #174.

Clarifies the cardinality of Script Event Description.

Updates the data model diagram (in a way that is non-ideal but hard to fix automatically: it may be worth replacing the PlantUML-generated diagram with a more manually constructed one prior to final publication), and defines Description Type more explicitly.

Also clarify that there is no uniqueness constraint for `daptm:descType` and that `xml:lang` can be used to label the language of the contents of a `ttm:desc`.