Change Primary Language to Default Language, add Original Language
Many editorial tweaks to ensure that we are not using the term "original language" in a way that could be confusing in comparison to "Original" or the newly defined "Original Language".

This change is intended to clarify the situation when content without an inherent textual language is transcribed, and what language it should be transcribed as.