Redraft opening section of §2.1 (#183)
Thanks for all the review inputs, merging.

* Redraft opening section of §2.1

Closes #175.

* Tweak language

Remove "English", and clarify that we are using domain-specific terms in place of the general usage.

* Attempt to address feedback

Clarify that a transcript is a type of script in general usage.

* Remove standalone use of "script"

Always clarify whether we mean transcript, or script, or both. Changes only required in the Introduction to achieve this.