Editorial improvements
Move the XML constraints of `daptm:langSrc` into a bullet list for consistency with other sections defining such lists of constraints.
Simplify the inclusion of `daptm:langSrc` in Text objects, and make it a SHOULD with an explanation of the feasibility and risks of omitting it in favour of inheritance.
Widen the scope of the `#textLanguageSource` feature to all elements.