Rework audio description original and translation languages (#179)
* Rework audio description original and translation languages

Closes #173 and #148.

Text Language Source updated to remove the link between it and Primary Language. Needed to introduce a new concept of a "nominal" original language for the programme, for the Original language of the Audio Description.

Added notes and examples about how to handle non-nominal original language in-image text in audio descriptions, and the potential need to translate it.

Tidied some of the existing notes.

* Change Primary Language to Default Language, add Original Language

Many editorial tweaks to ensure that we are not using the term "original language" in a way that could be confusing in comparison to "Original" or the newly defined "Original Language".

This change is intended to clarify the situation when content without an inherent textual language is transcribed, and what language it should be transcribed as.

* Remove notes about nominal language

* Add `#originalLang-root` extension and require it

* Change `daptm:langSrc` to use a BCP47 language code

Resolves #148 as per TTWG discussion.

* Make Text Language Source optional

* Fix `langSrc` in example

* Address review feedback

* change `daptm:originalLang` to a list `daptm:originalLangs` so that multiple language content can be properly represented
* fix up all the examples to match
* Make Original Language optional
* Make Text Language Source optional, and define a semantic for when it is omitted
* Fix up styling to remove the strange small text for notes and examples (other than examples that are code)
* Clarify how to represent transcript text with an inherent language vs without
* Remove the complex dependency on "Original Language" for text without an inherent language

* Address @cconcolato review feedback

* as -> since

* Address @cconcolato review feedback

* Fully specify `daptm:langSrc` including inheritance, allow empty string

* Editorial improvements

Move the XML constraints of `daptm:langSrc` into a bullet list for consistency with other sections defining such lists of constraints.
Simplify the inclusion of `daptm:langSrc` in Text objects, and make it a SHOULD with an explanation of the feasibility and risks of omitting it in favour of inheritance.
Widen the scope of the `#textLanguageSource` feature to all elements.

* Explicitly state the computed value is the specified value if present

* Address @cconcolato 's feedback

* Make SVG valid

* remove codeline attributes that got away last time

* Address some feedback

* Simplify note about As-Recorded scripts

* Extend `#xmlLang-root` to include Default Language requirements

i.e. that xml:lang cannot be empty.

* Delete Original Languages

Removes `daptm:originalLangs` and `#originalLangs-root` too.

* Simplify Script Event requirement for zero or more text objects

* Reorder bullets and para in Text Language Source

* Explain how to identify if Text is Original or Translation.

* Simplify language about "marking" translations

* Target implementers as well as authors in warning notes

* Address editorial feedback comments from @cconcolato

* Add note as suggested by @andreastai