Explicitly permit `daptm:represents` on tt, body, div, p and span
Closes #262.

Explicitly state the previous implied permission to put `daptm:represents` on content elements.

Update the XSD to reflect this permission (doesn't make any practical difference since 'other' metadata was already permitted, but it is clearer to see).

Add the attribute to each of the elements in the schema validator test fixture to verify that it continues to permit it and passes the test.