Explicitly permit daptm:represents on tt, body, div, p and span (#300)
* Explicitly permit `daptm:represents` on tt, body, div, p and span

Closes #262.

Explicitly state the previous implied permission to put `daptm:represents` on content elements.

Update the XSD to reflect this permission (doesn't make any practical difference since 'other' metadata was already permitted, but it is clearer to see).

Add the attribute to each of the elements in the schema validator test fixture to verify that it continues to permit it and passes the test.

* `daptm:langSrc` may also be specified on `<body>` and `<div>`

* Add table of attributes that can be specified on other elements

* Add Represents to Text objects

Includes #294, and closes both #262 and #295 by adding Represents to Text, both in the class diagram and in the data model descriptions of Text and Represents.

Includes a note advising that where possible, script events should be split if they represent different things.

* Update Using computed attribute values table

to add `<p>` and `<span>` for `daptm:represents`

* Simplify note text

Co-authored-by: Cyril Concolato <cconcolato@netflix.com>

* Update index.html

Co-authored-by: Cyril Concolato <cconcolato@netflix.com>

* Apply suggestions from code review

Co-authored-by: Cyril Concolato <cconcolato@netflix.com>

---------

Co-authored-by: Cyril Concolato <cconcolato@netflix.com>